requirements.txt
crewai
python-dotenv
crewai[tools]